Refactored recommendations by splitting into service and controller, deleted useless variables, and adjusted the tests
