+++ src/main/java/com/weatherAPI/weatherApplication/service/BikeService.java
@@ -11,8 +11,6 @@ import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import java.util.List;
-import java.util.ArrayList;
 import java.util.HashMap;
 
 @Service
